TSConfig Type Checking
This is the command I set up to demo this as well.

`tsc --noEmit --skipLibCheck --target esnext --module nodenext ./**/*.ts`

I have a 'template' config that I tend to start with, then I remove flags to make it on the same level as what the project needs, then bring flags back into place one by one, incrementally allowing for stronger type checking. :)

So right now this simply allows you to run `tsc` on the project, and you get a list of all current errors and issues to focus on, without having to go to each file to find them. It helps for migrations and larger cross-project updates as well.